+++ core/test/sort_test.cpp
@@ -10,13 +10,13 @@ using namespace scipp::core;
 
 TEST(SortTest, variable_1d) {
   const auto var =
-      makeVariable<int>(Dims{Dim::X}, Shape{3}, units::Unit(units::m),
-                        Values{1, 2, 3}, Variances{4, 5, 6});
+      makeVariable<float>(Dims{Dim::X}, Shape{3}, units::Unit(units::m),
+                          Values{1, 2, 3}, Variances{4, 5, 6});
   const auto key =
       makeVariable<int>(Dims{Dim::X}, Shape{3}, Values{10, 20, -1});
   const auto expected =
-      makeVariable<int>(Dims{Dim::X}, Shape{3}, units::Unit(units::m),
-                        Values{3, 1, 2}, Variances{6, 4, 5});
+      makeVariable<float>(Dims{Dim::X}, Shape{3}, units::Unit(units::m),
+                          Values{3, 1, 2}, Variances{6, 4, 5});
 
   EXPECT_EQ(sort(var, key), expected);
 }
@@ -44,8 +44,8 @@ TEST(SortTest, variable_2d) {
 TEST(SortTest, dataset_1d) {
   Dataset d;
   d.setData("a",
-            makeVariable<int>(Dims{Dim::X}, Shape{3}, units::Unit(units::m),
-                              Values{1, 2, 3}, Variances{4, 5, 6}));
+            makeVariable<float>(Dims{Dim::X}, Shape{3}, units::Unit(units::m),
+                                Values{1, 2, 3}, Variances{4, 5, 6}));
   d.setData("b",
             makeVariable<double>(Dims{Dim::X}, Shape{3}, units::Unit(units::s),
                                  Values{0.1, 0.2, 0.3}));
@@ -55,9 +55,9 @@ TEST(SortTest, dataset_1d) {
                                   Values{1, 2, 3}));
 
   Dataset expected;
-  expected.setData("a", makeVariable<int>(Dims{Dim::X}, Shape{3},
-                                          units::Unit(units::m),
-                                          Values{3, 1, 2}, Variances{6, 4, 5}));
+  expected.setData(
+      "a", makeVariable<float>(Dims{Dim::X}, Shape{3}, units::Unit(units::m),
+                               Values{3, 1, 2}, Variances{6, 4, 5}));
   expected.setData("b", makeVariable<double>(Dims{Dim::X}, Shape{3},
                                              units::Unit(units::s),
                                              Values{0.3, 0.1, 0.2}));
